新增-comments功能

+++ controllers/posts.js
@@ -1,4 +1,5 @@
 const Post = require("../models/postsModel");
+const Comment = require("../models/commentsModel");
 // const User = require("../models/usersModel");
 const handleErrorAsync = require("../service/handleErrorAsync");
 // const handleError = require("../service/handleError");
@@ -16,21 +17,30 @@ const posts = {
         path: "user",
         select: "name photo",
       })
+      .populate({
+        path: "comments",
+        select: "comment user",
+      })
       .sort(timeSort);
     res.status(200).json({
       status: "success",
       data: posts,
     });
   },
-  getPostById: async (req, res, next) => {
+  getPostById: handleErrorAsync(async (req, res, next) => {
     const id = req.params.id;
     if (!mongoose.isObjectIdOrHexString(id)) {
-      return next(appError(400, "查無此id", next));
+      return next(appError(400, "無效id", next));
     }
-    const post = await Post.findById(id).populate({
-      path: "user",
-      select: "name photo",
-    });
+    const post = await Post.findById(id)
+      .populate({
+        path: "user",
+        select: "name photo",
+      })
+      .populate({
+        path: "comments",
+        select: "comment user",
+      });
     if (post !== null) {
       res.status(200).json({
         status: "success",
@@ -39,7 +49,7 @@ const posts = {
     } else {
       return next(appError(400, "查無此id", next));
     }
-  },
+  }),
   createdPost: handleErrorAsync(async (req, res, next) => {
     const user = req.user.id; //form isAuth(token)
     const { content, image } = req.body;
@@ -58,15 +68,25 @@ const posts = {
     });
   }),
   updatedPost: handleErrorAsync(async (req, res, next) => {
+    const user = req.user.id; //form isAuth(token)
     const id = req.params.id;
     const { content, image } = req.body;
 
     if (!mongoose.isObjectIdOrHexString(id)) {
-      return next(appError(400, "查無此id", next));
+      return next(appError(400, "無效id", next));
     }
-
     if (!content || validator.isEmpty(content)) {
       return next(appError(400, "你沒有填寫 content 資料", next));
+    }
+    // check post exist
+    const curPost = await Post.findById(id);
+    if (curPost === null) {
+      return next(appError(400, "查無此id", next));
+    }
+    // check post create user id = update user id
+    const curCommentUserId = curPost.user;
+    if (curCommentUserId !== user) {
+      return next(appError(400, "無編輯權限", next));
     }
     const newPost = await Post.findByIdAndUpdate(
       id,
@@ -76,29 +96,33 @@ const posts = {
       },
       { new: true }
     );
-    if (newPost !== null) {
-      res.status(200).json({
-        status: "success",
-        data: newPost,
-      });
-    } else {
-      return next(appError(400, "查無此id", next));
-    }
+    res.status(200).json({
+      status: "success",
+      data: newPost,
+    });
   }),
   deletedPostById: handleErrorAsync(async (req, res, next) => {
+    const user = req.user.id;
     const id = req.params.id;
     if (!mongoose.isObjectIdOrHexString(id)) {
+      return next(appError(400, "無效id", next));
+    }
+    // check post exist
+    const curPost = await Post.findById(id);
+    if (curPost === null) {
       return next(appError(400, "查無此id", next));
     }
+    // check post create user id = update user id
+    const curCommentUserId = curPost.user;
+    if (curCommentUserId !== user) {
+      return next(appError(400, "無刪除權限", next));
+    }
+
     const delPost = await Post.findByIdAndDelete(id);
-    if (delPost !== null) {
-      res.status(200).json({
-        status: "success",
-        data: delPost,
-      });
-    } else {
-      return next(appError(400, "查無此id", next));
-    }
+    res.status(200).json({
+      status: "success",
+      data: delPost,
+    });
   }),
   clearPosts: async (req, res, next) => {
     await Post.deleteMany({});
@@ -112,7 +136,7 @@ const posts = {
     const user = req.user.id; //form isAuth(token)
 
     if (!mongoose.isObjectIdOrHexString(id)) {
-      return next(appError(400, "查無此id", next));
+      return next(appError(400, "無效id", next));
     }
     const newPost = await Post.findByIdAndUpdate(
       id,
@@ -136,7 +160,7 @@ const posts = {
     const user = req.user.id; //form isAuth(token)
 
     if (!mongoose.isObjectIdOrHexString(id)) {
-      return next(appError(400, "查無此id", next));
+      return next(appError(400, "無效id", next));
     }
     const newPost = await Post.findByIdAndUpdate(
       id,
@@ -155,15 +179,15 @@ const posts = {
       return next(appError(400, "查無此id", next));
     }
   }),
-  getPostsByUserId: async (req, res, next) => {
+  getPostsByUserId: handleErrorAsync(async (req, res, next) => {
     const userId = req.params.id;
     if (!mongoose.isObjectIdOrHexString(userId)) {
-      return next(appError(400, "查無此會員", next));
+      return next(appError(400, "無效id", next));
     }
 
     const posts = await Post.find({ user: userId }).populate({
-      path: "user",
-      select: "name photo",
+      path: "comments",
+      select: "comment user",
     });
 
     res.status(200).json({
@@ -171,7 +195,67 @@ const posts = {
       results: posts.length,
       posts,
     });
-  },
+  }),
+  addComment: handleErrorAsync(async (req, res, next) => {
+    const user = req.user.id; //form isAuth(token)
+    const post = req.params.id;
+    const { comment } = req.body;
+
+    if (!mongoose.isObjectIdOrHexString(post)) {
+      return next(appError(400, "無效id", next));
+    }
+
+    if (!comment || validator.isEmpty(comment)) {
+      return next(appError(400, "你沒有填寫 comment 資料", next));
+    }
+
+    const newComment = await Comment.create({
+      post,
+      user,
+      comment,
+    });
+    res.status(201).json({
+      status: "success",
+      data: {
+        comments: newComment,
+      },
+    });
+  }),
+  updateComment: handleErrorAsync(async (req, res, next) => {
+    const user = req.user.id; //form isAuth(token)
+    const commentId = req.params.id;
+    const { comment } = req.body;
+
+    if (!mongoose.isObjectIdOrHexString(commentId)) {
+      return next(appError(400, "無效id", next));
+    }
+    if (!comment || validator.isEmpty(comment)) {
+      return next(appError(400, "你沒有填寫 comment 資料", next));
+    }
+    // check comment exist
+    const curComment = await Comment.findById(commentId);
+    if (curComment === null) {
+      return next(appError(400, "查無此id", next));
+    }
+    // check comment create user id === update user id
+    const curCommentUserId = curComment.user.id;
+    if (curCommentUserId !== user) {
+      return next(appError(400, "無編輯權限", next));
+    }
+    const newComment = await Comment.findByIdAndUpdate(
+      commentId,
+      {
+        comment,
+      },
+      { new: true }
+    );
+    res.status(201).json({
+      status: "success",
+      data: {
+        comments: newComment,
+      },
+    });
+  }),
 };
 
 module.exports = posts;
